[BOOK-67] refactor: apis - 패키지 이동
- AuthCredentials은 request에서 변환되는 중간 데이터 dto이기에 dto 패키지 내부로 이동
