+++ apis/src/main/kotlin/org/yapp/apis/auth/service/SocialAuthService.kt
@@ -1,6 +1,7 @@
 package org.yapp.apis.auth.service
 
 import org.springframework.stereotype.Service
+import org.yapp.apis.auth.dto.AuthCredentials
 import org.yapp.apis.auth.exception.AuthErrorCode
 import org.yapp.apis.auth.exception.AuthException
 import org.yapp.apis.auth.strategy.AuthStrategy
