+++ apis/src/main/kotlin/org/yapp/apis/auth/usecase/AuthUseCase.kt
@@ -3,7 +3,7 @@ package org.yapp.apis.auth.usecase
 import org.springframework.stereotype.Service
 import org.yapp.apis.auth.dto.response.TokenPairResponse
 import org.yapp.apis.auth.dto.response.UserProfileResponse
-import org.yapp.apis.auth.service.AuthCredentials
+import org.yapp.apis.auth.dto.AuthCredentials
 import org.yapp.apis.auth.service.SocialAuthService
 import org.yapp.apis.auth.service.TokenService
 import org.yapp.apis.auth.service.UserAuthService
